move choice attributes from listener to type
